Rename variable and files
